refactor : 테스트 및 기능 리팩토링
