+++ src/test/java/subway/common/SectionRestAssuredCRUD.java
@@ -1,0 +1,30 @@
+package subway.common;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.springframework.http.MediaType;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class SectionRestAssuredCRUD {
+
+    public static ExtractableResponse<Response> addSection(Long upStationId, Long downStationId, int distance, Long lineNum) {
+
+        Map<String, Object> paraMap = new HashMap<>();
+        paraMap.put("upStationId", upStationId);
+        paraMap.put("downStationId", downStationId);
+        paraMap.put("distance", 10);
+
+        return RestAssured
+                .given().log().all()
+                    .contentType(MediaType.APPLICATION_JSON_VALUE)
+                    .pathParam("id", lineNum)
+                    .body(paraMap)
+                .when()
+                    .post("/lines/{id}/sections")
+                .then().log().all()
+                .extract();
+    }
+}
